+++ R/xml_structure.R
@@ -7,6 +7,7 @@
 #'
 #' @param x HTML/XML document (or part there of)
 #' @param indent Number of spaces to ident
+#' @inheritParams cat
 #' @export
 #' @examples
 #' xml_structure(read_xml("<a><b><c/><c/></b><d/></a>"))
@@ -17,43 +18,45 @@
 #'
 #' h <- read_html("<body><p id = 'a'></p><p class = 'c d'></p></body>")
 #' html_structure(h)
-xml_structure <- function(x, indent = 2) {
-  tree_structure(x, indent = indent, html = FALSE)
+xml_structure <- function(x, indent = 2, file = "") {
+  cat(file = file)
+  tree_structure(x, indent = indent, html = FALSE, file = file)
 }
 
 #' @export
 #' @rdname xml_structure
-html_structure <- function(x, indent = 2) {
-  tree_structure(x, indent = indent, html = TRUE)
+html_structure <- function(x, indent = 2, file = "") {
+  cat(file = file)
+  tree_structure(x, indent = indent, html = TRUE, file = file)
 }
 
-tree_structure <- function(x, indent = 2, html = FALSE) {
+tree_structure <- function(x, indent = 2, html = FALSE, file = "") {
   UseMethod("tree_structure")
 }
 
 #' @export
-tree_structure.xml_missing <- function(x, indent = 2, html = FALSE) {
+tree_structure.xml_missing <- function(x, indent = 2, html = FALSE, file = "") {
   NA_character_
 }
 
 #' @export
-tree_structure.xml_nodeset <- function(x, indent = 2, html = FALSE) {
+tree_structure.xml_nodeset <- function(x, indent = 2, html = FALSE, file = "") {
   for (i in seq_along(x)) {
-    cat("[[", i, "]]\n", sep = "")
-    print_xml_structure(x[[i]], indent = indent, html = html)
-    cat("\n")
+    cat("[[", i, "]]\n", sep = "", file = file, append = TRUE)
+    print_xml_structure(x[[i]], indent = indent, html = html, file = file)
+    cat("\n", file = file, append = TRUE)
   }
 
   invisible()
 }
 
 #' @export
-tree_structure.xml_node <-  function(x, indent = 2, html = FALSE) {
-  print_xml_structure(x, indent = indent, html = html)
+tree_structure.xml_node <-  function(x, indent = 2, html = FALSE, file = "") {
+  print_xml_structure(x, indent = indent, html = html, file = file)
   invisible()
 }
 
-print_xml_structure <- function(x, prefix = 0, indent = 2, html = FALSE) {
+print_xml_structure <- function(x, prefix = 0, indent = 2, html = FALSE, file = "") {
   padding <- paste(rep(" ", prefix), collapse = "")
   type <- xml_type(x)
 
@@ -83,10 +86,10 @@ print_xml_structure <- function(x, prefix = 0, indent = 2, html = FALSE) {
 
     node <- paste0("<", xml_name(x), attr_str, ">")
 
-    cat(padding, node, "\n", sep = "")
+    cat(padding, node, "\n", sep = "", file = file, append = TRUE)
     lapply(xml_contents(x), print_xml_structure, prefix = prefix + indent,
-      indent = indent, html = html)
+      indent = indent, html = html, file = file)
   } else {
-    cat(padding, "{", type, "}\n", sep = "")
+    cat(padding, "{", type, "}\n", sep = "", file = file, append = TRUE)
   }
 }