+++ revdep/check.R
@@ -1,5 +1,0 @@
-library("devtools")
-
-res <- revdep_check()
-revdep_check_save_summary()
-revdep_check_print_problems()