+++ NEWS.md
@@ -1,5 +1,8 @@
 # xml2 1.2.0.9000
 
+* print methods now match the type of document, e.g. `read_html()` prints as
+  "{html_document}" rather than "{xml_document}" (#227)
+
 * Test failure fixed due to behavior change with relative paths in libxml2
   2.9.9 (#245).