+++ NEWS.md
@@ -1,3 +1,5 @@
+# xml2 (development version)
+
 # xml2 1.2.1
 
 ## New Features