+++ NEWS.md
@@ -1,4 +1,7 @@
-# xml2 (development version)
+# xml2 1.2.2
+
+* Export S4 classes with documentation, so they can be used in other packages
+without Warnings (@nuest, #267)
 
 # xml2 1.2.1